Updated UserDTO, article business logic

+++ BACKEND/girls_for_girls/src/main/java/com/neobis/g4g/girls_for_girls/repository/UserGroupRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface UserGroupRepository extends JpaRepository<UserGroup, Integer> {
 
-    UserGroup findById(int id);
+    UserGroup findById(long id);
 }
